[WIP] patches: Hold vGaming-related patches
This patch attempted to add a (custom, consumer-based) vGaming profile
for Turing. However, we don't support vGaming anywhere else. Let's hold off
on this for now.

Another, possibily related? patch attempted to add direct support for
the GeForce RTX 2070 as a vGPU-capable card, but this card is already
mapped to the relevant Quadro. So, drop it. It can be reinstated later
if that's wanted.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>